+++ lib/sparql/query/result/solution_mapping.ex
@@ -1,0 +1,31 @@
+defmodule SPARQL.Query.Result.SolutionMapping do
+  @moduledoc !"""
+  Some shared functions on solution mappings for internal purposes.
+  """
+
+  @doc """
+  Two solution mappings `m1` and `m2` are compatible if, for every variable `v in `dom(m1)` and in `dom(m2)`, `m1(v) = m2(v)`.
+
+  see Definition in <https://www.w3.org/TR/sparql11-query/#BasicGraphPattern>
+  """
+  def compatible?(m1, m2) do
+    Enum.all? m1, fn
+      {:__id__, _} ->
+        true
+
+      {k, v} ->
+        Map.get(m2, k, v) == v
+    end
+  end
+
+  @doc """
+  Merges two solution mappings.
+
+  see Definition in <https://www.w3.org/TR/sparql11-query/#BasicGraphPattern>
+  """
+  def merge(m1, m2) do
+    m1
+    |> Map.merge(m2)
+    |> Map.put(:__id__, make_ref())
+  end
+end